+++ CHANGELOG.md
@@ -6,7 +6,7 @@ AdminNeo 5.5.0
 
 ### Changes
 
-
+- Check max_file_uploads in SQL import form
 
 ### Bugfixes